extracts json utils to utilities common folder
