+++ app/javascript/packs/custom_policy.js
@@ -1,7 +1,7 @@
 import {render} from 'react-dom'
 import React from 'react'
 import {CustomPolicy} from 'Policies/components/CustomPolicy'
-import {safeFromJsonString} from 'Policies/util'
+import {safeFromJsonString} from 'utilities/json-utils'
 
 document.addEventListener('DOMContentLoaded', () => {
   const container = document.getElementById('custom-policy-container')
